+++ CHANGELOG.markdown
@@ -3,6 +3,9 @@
 
 ## Version 0.9.1 - *Not Yet Released*
 
+ * Added a custom Date#strftime extension to be used for formatting dates with
+   format strings.
+
  * Added "enabled" state management to Aphid.UI.View through the isEnabled
    property and the enable() and disable() methods.